classes for introducing thought into tool calling llm
